fix: Add src directory to path so test dir can access src files; refactor step dir names to avoid path smashing.
