--- test/steps/__init__.py
@@ -1,2 +1,0 @@
-# from test import *
-from src import *
